refactor: 접근 제한자 및 변수 순서 컨벤션에 맞게 수정
