+++ tomcat/src/main/java/org/apache/coyote/http11/handler/IndexHandler.java
@@ -1,6 +1,6 @@
 package org.apache.coyote.http11.handler;
 
-import org.apache.coyote.http11.ResourceUtil;
+import org.apache.coyote.http11.handler.support.FileReader;
 import org.apache.coyote.http11.model.ContentType;
 import org.apache.coyote.http11.model.HttpRequest;
 import org.apache.coyote.http11.model.HttpResponse;
@@ -11,7 +11,7 @@ public class IndexHandler implements Handler {
 
     @Override
     public String getResponse() {
-        String responseBody = ResourceUtil.getResponseBody("/index.html", getClass());
+        String responseBody = FileReader.getFile("/index.html", getClass());
 
         HttpResponse httpResponse = HttpResponse.from(ContentType.HTML, responseBody);
         return httpResponse.getResponse();
