+++ tomcat/src/main/java/org/apache/coyote/http11/handler/ResourceHandler.java
@@ -1,22 +1,22 @@
 package org.apache.coyote.http11.handler;
 
-import org.apache.coyote.http11.ResourceUtil;
+import org.apache.coyote.http11.handler.support.FileReader;
 import org.apache.coyote.http11.model.ContentType;
 import org.apache.coyote.http11.model.HttpRequest;
 import org.apache.coyote.http11.model.HttpResponse;
 
-public class FileHandle implements Handler{
+public class ResourceHandler implements Handler{
 
     private final HttpRequest httpRequest;
 
-    public FileHandle(final HttpRequest httpRequest) {
+    public ResourceHandler(final HttpRequest httpRequest) {
         this.httpRequest = httpRequest;
     }
 
     @Override
     public String getResponse() {
         String path = httpRequest.getPath();
-        String responseBody = ResourceUtil.getResponseBody(path, getClass());
+        String responseBody = FileReader.getFile(path, getClass());
 
         ContentType contentType = ContentType.of(path);
         HttpResponse httpResponse = HttpResponse.from(contentType, responseBody);
