+++ tomcat/src/main/java/org/apache/coyote/http11/handler/support/FileReader.java
@@ -1,18 +1,18 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.handler.support;
 
 import java.net.URL;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 
-public class ResourceUtil {
+public class FileReader {
 
-    public static String getResponseBody(final String path, final Class<?> classes) {
+    public static String getFile(final String path, final Class<?> classes) {
         try {
             URL resource = classes.getResource("/static" + path);
             return Files.readString(Paths.get(resource.toURI()), StandardCharsets.UTF_8);
         } catch (Exception exception) {
-            return "파일을 불러올 수 없습니다.";
+            throw new IllegalArgumentException("파일을 불러올 수 없습니다.");
         }
     }
 }
