+++ tomcat/src/main/java/org/apache/coyote/http11/model/ContentType.java
@@ -15,7 +15,7 @@ public enum ContentType {
         this.extension = extension;
     }
 
-    public static ContentType of(String input) {
+    public static ContentType of(final String input) {
         return Arrays.stream(values())
                 .filter(i -> input.contains(i.extension))
                 .findAny()
